Server/requirements.txt
grpcio>=1.62.0
grpcio-tools>=1.62.0
pika>=1.3.2